@@ -2,10 +2,10 @@ name: examples
 description: dart.dev example code.
 
 environment:
-  sdk: ^3.0.0
+  sdk: ^3.1.0
 
 dependencies:
   examples_util: {path: ../util}
 
 dev_dependencies:
-  lints: ^2.1.0
+  lints: ^3.0.0
