@@ -2,7 +2,7 @@ name: examples
 description: dart.dev example code.
 
 environment:
-  sdk: ^3.0.0
+  sdk: ^3.1.0
 
 dev_dependencies:
-  lints: ^2.1.0
+  lints: ^3.0.0
