@@ -2,10 +2,10 @@ name: build_runner_usage
 description: dart.dev build_runner example code.
 
 environment:
-  sdk: ^3.0.0
+  sdk: ^3.1.0
 
 dev_dependencies:
   args: ^2.4.0
-  build_runner: ^2.4.1
-  build_test: ^2.1.7
-  lints: ^2.1.0
+  build_runner: ^2.4.6
+  build_test: ^2.2.0
+  lints: ^3.0.0
