@@ -3,8 +3,8 @@ description: dart.dev example code.
 version: 1.0.0
 
 environment:
-  sdk: ^3.0.0
+  sdk: ^3.1.0
 
 dev_dependencies:
-  lints: ^2.1.0
-  test: ^1.24.2
+  lints: ^3.0.0
+  test: ^1.24.6
