@@ -4,11 +4,11 @@ version: 0.0.1
 publish_to: none
 
 environment:
-  sdk: ^3.0.0
+  sdk: ^3.1.0
 
 dependencies:
   examples_util: {path: ../util}
 
 dev_dependencies:
-  lints: ^2.1.0
-  test: ^1.24.2
+  lints: ^3.0.0
+  test: ^1.24.6
