@@ -3,12 +3,9 @@ description: dart.dev example code.
 version: 0.0.1
 
 environment:
-  sdk: ^3.0.0
-
-dependencies:
-  html: any
+  sdk: ^3.1.0
 
 dev_dependencies:
-  test: ^1.24.2
-  lints: ^2.1.0
+  lints: ^3.0.0
+  test: ^1.24.6
   examples_util: {path: ../util}
