@@ -1,3 +1,3 @@
 import 'package:dart_sdk_archive/dart_sdk_archive.dart' as sdk_archive_app;
 
-void main() => sdk_archive_app.runApp();
+void main() async => await sdk_archive_app.runApp();
