@@ -1,5 +1,3 @@
-library sdk_builds.version_info;
-
 import 'package:pub_semver/pub_semver.dart';
 
 final _oldRevisionPostfix = RegExp(r'(\d+\.\d+\.\d+)\.(\d+)_r(\d+)');
@@ -11,12 +9,20 @@ abstract class VersionInfo implements Comparable<VersionInfo> {
   final String revisionPath;
   final DateTime? creationTime;
 
-  VersionInfo(this.version, this.date, this.channel, this.revisionPath,
-      {this.creationTime});
+  VersionInfo(
+    this.version,
+    this.date,
+    this.channel,
+    this.revisionPath, {
+    this.creationTime,
+  });
 
   static VersionInfo parse(
-      String channel, String revisionPath, Map<String, dynamic> json,
-      {DateTime? creationTime}) {
+    String channel,
+    String revisionPath,
+    Map<String, Object?> json, {
+    DateTime? creationTime,
+  }) {
     // Date parse magic
     var dateJson = json['date'] as String;
     DateTime date;
@@ -44,12 +50,24 @@ abstract class VersionInfo implements Comparable<VersionInfo> {
       // assume git!
       assert(revision.length == 40);
 
-      return GitVersionInfo(version, date, channel, revisionPath, revision,
-          creationTime: creationTime);
+      return GitVersionInfo(
+        version,
+        date,
+        channel,
+        revisionPath,
+        revision,
+        creationTime: creationTime,
+      );
     }
 
-    return SvnVersionInfo(version, date, channel, revisionPath, svnRevision,
-        creationTime: creationTime);
+    return SvnVersionInfo(
+      version,
+      date,
+      channel,
+      revisionPath,
+      svnRevision,
+      creationTime: creationTime,
+    );
   }
 
   @override
@@ -62,15 +80,25 @@ abstract class VersionInfo implements Comparable<VersionInfo> {
 class SvnVersionInfo extends VersionInfo {
   final int revision;
 
-  SvnVersionInfo(super.version, super.date, super.channel, super.revisionPath,
-      this.revision,
-      {super.creationTime});
+  SvnVersionInfo(
+    super.version,
+    super.date,
+    super.channel,
+    super.revisionPath,
+    this.revision, {
+    super.creationTime,
+  });
 }
 
 class GitVersionInfo extends VersionInfo {
   final String ref;
 
   GitVersionInfo(
-      super.version, super.date, super.channel, super.revisionPath, this.ref,
-      {super.creationTime});
+    super.version,
+    super.date,
+    super.channel,
+    super.revisionPath,
+    this.ref, {
+    super.creationTime,
+  });
 }
